Added a function that prints an integer.

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
  */
 void print_number(int n)
 {
-	short int num;
+	unsigned int num;
 
 	if (n < 0)
 	{
